pyproject.toml
[project]
name = "scranton-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "langchain-core>=1.2.5",
    "langgraph>=1.0.5",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
